requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
azure-ai-textanalytics